Add bloom filter
When use find option, Bloom will process first.
If the word isn't found by Bloom filter, we can say that word isn't
in the list.

Now if user want to add a string to tree, program will check bloom
filter to see if there the same string in the tree.

It will make performance increased because if the string is already
in the tree, it does not need to walk through tree.