Refactor the operations retrieving next node (#5)
New function "tst_ins" and "tst_del" are decoupled from function
"tst_ins_del", reducing the amount of comparisons.

If the root of tst is NULL, "pcurr" will be pointed to it when processed
in the loop. The root node will be allocated for external memory at the
first glance, and then the comparison will never be true.